requirements.txt
ipykernel==7.0.0a1
jupyterlab==4.5.0a1
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.69
langchain-openai==0.3.28
langgraph==0.5.3
langgraph-api==0.2.98
langgraph-cli==0.3.5
langgraph-checkpoint-sqlite==2.0.10
pydantic==2.11.7
python-dotenv==1.1.1